Enable 'show_prev_next' in the documented defaults
Switch the default value of 'show_prev_next' to True in the
theme options documentation.

The reason for this change is that the prev / next buttons are
useful for navigating the documentation, especially in narrow
viewports.
In that case, the only other option to naviagate is the hamburger
menu, which the user may not be aware of.